+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v0.4.2
 * Added Json export
+* Sort cards and other items in the per mod export
 
 ### v0.4.1
 * Fix file extension of creature images (#2)